+++ OpenGL/src/VertexBuffer.h
@@ -6,7 +6,7 @@
 class VertexBuffer
 {
 public:
-	VertexBuffer(const void* data, unsigned int size);
+	VertexBuffer(const void* data, size_t size);
 	~VertexBuffer();
 	void bind() const;
 	void unbind() const;
